+++ frontend/src/App.js
@@ -1,6 +1,8 @@
 import React, { useEffect } from "react";
 import AllRoute from "./Route/AllRoute";
 import { injectStyle } from "react-toastify/dist/inject-style";
+import { ToastContainer } from "react-toastify";
+
 function App() {
   useEffect(() => {
     injectStyle();
@@ -8,6 +10,12 @@ function App() {
   return (
     <>
       <AllRoute />
+      <ToastContainer
+        autoClose={2000}
+        hideProgressBar
+        pauseOnHover={false}
+        closeOnClick
+      />
     </>
   );
 }
