+++ frontend/src/Route/AllRoute.js
@@ -4,11 +4,11 @@ import RegisterPage from "../components/Register/RegisterPage";
 import { Routes, Route, BrowserRouter } from "react-router-dom";
 import SignUpPage from "../components/SignUp/SignUpPage";
 import MyPostsPage from "../components/MyPosts/MyPostsPage";
-import Setting from "../components/Setting/Setting";
 import StudyPage from "../components/Study/StudyPage";
 import KakaoSocial from "../components/Kakao/KakaoSocial";
 import KakaoSignUp from "../components/Kakao/KakaoSignUp";
 import GoogleSignUp from "../components/Google/GoogleSignUp";
+import SettingPage from "../components/Setting/SettingPage";
 
 function AllRoute() {
   return (
@@ -18,7 +18,7 @@ function AllRoute() {
         <Route path="/register" element={<RegisterPage />} />
         <Route path="/signup" element={<SignUpPage />} />
         <Route path="/myPosts" element={<MyPostsPage />} />
-        <Route path="/setting" element={<Setting />} />
+        <Route path="/setting" element={<SettingPage />} />
         <Route path="/study/:id" element={<StudyPage />} />
         <Route path="/oauth/callback/kakao" element={<KakaoSocial />} />
         <Route path="/kakaoSignup" element={<KakaoSignUp />} />
