+++ frontend/src/components/Main/MainBanner.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Banner } from "../components";
+import { Banner } from "./mainComponents";
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
