+++ frontend/src/util/deleteMultipleCookies.js
@@ -1,0 +1,6 @@
+import { deleteCookie } from "./cookie";
+
+export default function deleteMultipleCookies() {
+  deleteCookie("jwtToken");
+  deleteCookie("refreshToken");
+}
