+++ frontend/src/util/logout.js
@@ -1,16 +1,14 @@
-import { deleteCookie } from "./cookie";
-import { setCurrentPost, fetchUser } from "../slice/userSlice";
+import { setCurrentPost, setUser } from "../slice/userSlice";
 import { toast } from "react-toastify";
+import deleteMultipleCookies from "./deleteMultipleCookies";
 
 export const logout = (dispatch, navigate) => {
-  navigate("/");
+  dispatch(setUser([]));
   document.cookie = "jwtToken=; expires=Thu, 01 Jan 1970 00:00:01 GMT; path=/;";
   document.cookie =
     "refreshToken=; expires=Thu, 01 Jan 1970 00:00:01 GMT; path=/;";
-  deleteCookie("jwtToken");
-  deleteCookie("refreshToken");
+  deleteMultipleCookies();
   dispatch(setCurrentPost({}));
-  dispatch(fetchUser(""));
-  toast.error("로그아웃되었습니다.");
+  toast.success("로그아웃되었습니다.");
   navigate("/");
 };
