+++ frontend/src/util/registerOption.js
@@ -1,0 +1,38 @@
+export const registerOption = {
+  category: [
+    { value: "스터디", label: "스터디" },
+    { value: "프로젝트", label: "프로젝트" },
+  ],
+  people: [
+    { value: "인원 미정", label: "인원 미정" },
+    { value: "1명", label: "1명" },
+    { value: "2명", label: "2명" },
+    { value: "3명", label: "3명" },
+    { value: "4명", label: "4명" },
+    { value: "5명", label: "5명" },
+    { value: "6명", label: "6명" },
+    { value: "7명", label: "7명" },
+    { value: "8명", label: "8명" },
+    { value: "9명", label: "9명" },
+    { value: "10명", label: "10명" },
+  ],
+  howTo: [
+    { value: "온라인", label: "온라인" },
+    { value: "오프라인", label: "오프라인" },
+  ],
+  contact: [
+    { value: "카카오톡 오픈채팅", label: "카카오톡 오픈채팅" },
+    { value: "이메일", label: "이메일" },
+    { value: "구글폼", label: "구글폼" },
+  ],
+  duration: [
+    { value: "기간 미정", label: "기간 미정" },
+    { value: "1개월", label: "1개월" },
+    { value: "2개월", label: "2개월" },
+    { value: "3개월", label: "3개월" },
+    { value: "4개월", label: "4개월" },
+    { value: "5개월", label: "5개월" },
+    { value: "6개월", label: "6개월" },
+    { value: "장기", label: "장기" },
+  ],
+};
