+++ frontend/src/util/settingMultipleCookie.js
@@ -1,0 +1,6 @@
+import { setCookie } from "./cookie";
+
+export default function settingMultipleCookie(access, refresh, option) {
+  setCookie("jwtToken", access, option);
+  setCookie("refreshToken", refresh, option);
+}
